240506_스프링 시큐리티 로그인 구현

+++ src/main/java/com/wook/toy/repository/MemberRepository.java
@@ -1,12 +1,13 @@
 package com.wook.toy.repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wook.toy.domain.Member;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, BigDecimal> {
 
 	Optional<Member> findByUserId(String userId);
 }
